agno-hello-tool/pyproject.toml
[project]
name = "agno-hello-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.5.16",
    "groq>=1.1.2",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
